+++ src/open_icu/callbacks/proto.py
@@ -1,8 +1,9 @@
-from typing import Any, Protocol
+from typing import Any, Protocol, runtime_checkable
 
-from polars import LazyFrame, Expr
+from polars import Expr, LazyFrame
 
 
+@runtime_checkable
 class CallbackProtocol(Protocol):
     def __init__(self, **kwargs: Any) -> None:
         ...
